moved some classes | rename domain to host
